+++ sale-page-et2-maxim-update.php
@@ -135,30 +135,6 @@
                         <div class="block-picture-primary-container">
 
                             <div class="block-container">
-                                <a href="<?php echo site_url('/brands/currey-and-company/'); ?>">
-                                    <div class="block-picture">
-                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/09/6000-0581_1_-scaled.jpg">', wp_get_attachment_metadata(175830), 175830 ); ?>
-                                    </div>
-                                    <div class="block-caption-wrapper">
-                                        <div class="block-caption-container">
-                                            <h4 class="block-heading">Currey and Co.</h4>
-                                            <h3 class="block-heading">Save on All Currey and Co. Products</h3>
-                                            <h4 class="block-heading">Prices already discounted!</h4>
-                                            <a style="text-align: center;" href="<?php echo site_url('/brands/currey-and-company/'); ?>" class="block-shop-link">Shop Now</a>
-                                        </div>
-                                    </div>
-                                </a>
-                            </div>
-
-                        </div>
-
-                    </div>
-
-                    <div class="col-md-4">
-
-                        <div class="block-picture-primary-container">
-
-                            <div class="block-container">
                                 <a href="<?php echo site_url('/brands/elk-home'); ?>">
                                     <div class="block-picture">
                                         <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/07/elk_home_sale_august.jpg">', wp_get_attachment_metadata(172838), 172838 ); ?>
@@ -193,30 +169,6 @@
                                             <h3 class="block-heading">Get 20% off on All Elk Lighting</h3>
                                             <h4 class="block-heading">Prices already discounted!</h4>
                                             <a style="text-align: center;" href="<?php echo site_url('/brands/elk-lighting'); ?>" class="block-shop-link">Shop Now</a>
-                                        </div>
-                                    </div>
-                                </a>
-                            </div>
-
-                        </div>
-
-                    </div>
-
-                    <div class="col-md-4">
-
-                        <div class="block-picture-primary-container">
-
-                            <div class="block-container">
-                                <a href="<?php echo site_url('/product-category/artcraft-lighting/artcraft-lighting-sale'); ?>">
-                                    <div class="block-picture">
-                                        <?php echo wp_image_add_srcset_and_sizes('<img style="height: auto; width: 100%;" src="/wp-content/uploads/2020/03/AC7604BB.jpg">', wp_get_attachment_metadata(163669), 163669 ); ?>
-                                    </div>
-                                    <div class="block-caption-wrapper">
-                                        <div class="block-caption-container">
-                                            <h4 class="block-heading">Artcraft Lighting</h4>
-                                            <h3 class="block-heading">Get 30% off on select Artcraft Lighting</h3>
-                                            <h4 class="block-heading">Prices already discounted!</h4>
-                                            <a style="text-align: center;" href="<?php echo site_url('/product-category/artcraft-lighting/artcraft-lighting-sale'); ?>" class="block-shop-link">Shop Now</a>
                                         </div>
                                     </div>
                                 </a>