+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Elk home sale on front page to coupon
 - Elk home and lighting to coupon on sales page
+- Home page sale banners
 
 ## Removed