+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [ released ] - 12.4.2020
 
+## [ 1.29.0 ]
+
 ## Added
 
 - Elk lighting sale coupon to single product pages